Move Hadoop FileSystem classes to trino-hdfs module
