+++ lib/trino-hdfs/pom.xml
@@ -24,6 +24,11 @@
 
         <dependency>
             <groupId>io.trino</groupId>
+            <artifactId>trino-memory-context</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>io.trino</groupId>
             <artifactId>trino-plugin-toolkit</artifactId>
         </dependency>
