+++ lib/trino-hdfs/src/main/java/io/trino/hdfs/MemoryAwareFileSystem.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.filesystem;
+package io.trino.hdfs;
 
 import io.trino.memory.context.AggregatedMemoryContext;
 import org.apache.hadoop.fs.Path;
